add mobile and tablet
